vimba: storing necessary vimba includes and *.so files (#1)
We are building now vision not only for Linux x64 but also for ARM x64.
I did not store the currently used Vimba SDK for other architectures so I'm
updating it from 4.2 >> 5.1 version for the ability of cross-compiling
with the same version.

This PR updates the Allied Vision Vimba library that we use for connecting
with their Mako G-158C camera. In case we would have to build/recover,
copies of vimba packages for different systems are stored on our google
drive: here

For our luck, the API did not change and is back compatible.